# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Karabakh Atlas API
  description: >
    REST API serving region/city data for the Karabakh Atlas map website.
    Region/city write endpoints (POST/PUT/DELETE) require a logged-in
    session with the "admin" role (see /auth/login). Anyone can register
    an account via /auth/register, but new accounts get the "user" role,
    which can read everything but not write regions/cities.
  version: "1.0"
servers:
  - url: /api/v1
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: >
        Session cookie set by POST /auth/login or /auth/register. The
        session user additionally needs the "admin" role for
        region/city write endpoints.
  schemas:
...
    ForumPost:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        author_id:
          type: string
          readOnly: true
        author_name:
          type: string
          readOnly: true
        author_tier:
          type: string
          nullable: true
          readOnly: true
          enum: [newcomer, contributor, local_expert]
          description: >
            The author's contribution tier (see User.contribution_tier),
            based on their total approved post count, not just this one.
        target_city_id:
          type: string
          nullable: true
          description: >
            The city/POI this opinion is about, or null for a general
            opinion about Karabakh as a whole.
        target_city_name:
          type: string
          nullable: true
          readOnly: true
        target_news_id:
          type: string
          nullable: true
          description: >
            Set (instead of target_city_id) when this post is a comment
            on a news item.
        target_news_title:
          type: string
          nullable: true
          readOnly: true
        topic:
          type: string
          nullable: true
          readOnly: true
          enum: [general, global, student_life, about_karabakh, events_holidays, introductions]
          description: >
            The discussion topic of a general post (older posts read as
            general). Null for place- and news-scoped posts.
        parent_id:
          type: string
          nullable: true
          description: >
            Set when this post is a reply: the top-level post it
            answers. Threads are one level deep, and a reply is hidden
            from public lists whenever its parent is not approved.
        parent_author_name:
          type: string
          nullable: true
          readOnly: true
        parent_excerpt:
          type: string
          nullable: true
          readOnly: true
          description: First 140 characters of the post this one answers.
        body:
          type: string
          maxLength: 2000
        status:
          type: string
          enum: [pending, approved, rejected]
          readOnly: true
        moderated_by:
          type: string
          nullable: true
          readOnly: true
        moderated_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    ForumPostInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          maxLength: 2000
        target_city_id:
          type: string
          nullable: true
        target_news_id:
          type: string
          nullable: true
          description: >
            Comment on this news item. Mutually exclusive with
            target_city_id.
        topic:
          type: string
          enum: [general, global, student_life, about_karabakh, events_holidays, introductions]
          description: >
            Topic for a general post (defaults to general). Rejected on
            place- or news-scoped posts.
        parent_id:
          type: string
          description: >
            Reply to this published post (answering a reply attaches to
            the top-level post). Starts pending like any post. A reply
            takes its city/news/topic from its parent, so sending
            target_city_id, target_news_id, or topic with it is a 400.
...
paths:
...
  /forum/posts:
    get:
      summary: List forum posts
      description: >
        Defaults to only "approved" posts. ?mine=true (requires login)
        returns the caller's own posts regardless of status. ?status=
        (pending/approved/rejected) is only honored for an admin
        caller — anyone else's request for it silently falls back to
        approved-only, so unmoderated content never leaks.
      parameters:
        - name: city_id
          in: query
          description: >
            Only posts about this city/POI. Omit for general
            (target_city_id is null) opinions about Karabakh overall.
          schema:
            type: string
        - name: news_id
          in: query
          description: >
            Only comments on this news item. Omit both city_id and
            news_id for general opinions about Karabakh overall.
          schema:
            type: string
        - name: topic
          in: query
          description: >
            Only general posts in this topic (older posts count as
            general).
          schema:
            type: string
            enum: [general, global, student_life, about_karabakh, events_holidays, introductions]
        - name: mine
          in: query
          schema:
            type: boolean
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: List of forum posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ForumPost"
        "401":
          description: mine=true without a session
    post:
      summary: Submit an opinion (always starts "pending")
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForumPostInput"
      responses:
        "201":
          description: Post created, pending moderation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumPost"
        "400":
          description: Validation error
        "401":
          description: Login required
        "429":
          description: Too many posts submitted recently — try again later
  /forum/posts/{post_id}:
    parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
    delete:
      summary: Delete a post (its own author, or any admin)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Deleted
        "401":
          description: Login required
        "403":
          description: Not allowed to delete this post
        "404":
          description: Not found
  /forum/posts/{post_id}/status:
    parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
    put:
      summary: Approve or reject a post (admin only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [approved, rejected]
      responses:
        "200":
          description: Post moderated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumPost"
        "400":
          description: Validation error
        "401":
          description: Login required
        "403":
          description: Admin privileges required
...